Fix types

+++ include/ANurbs/src/Knots.h
@@ -15,9 +15,8 @@ public:
         const TScalar& t)
     {
         size_t span = std::upper_bound(std::begin(knots) + degree,
-                          std::end(knots) - degree, t)
-            - std::begin(knots) - 1;
-        return span;
+            std::end(knots) - degree, t) - std::begin(knots) - 1;
+        return static_cast<int>(span);
     }
 
     template <typename TScalar, typename TKnots>
@@ -28,9 +27,8 @@ public:
         const TScalar& t)
     {
         size_t span = std::lower_bound(std::begin(knots) + degree,
-                          std::end(knots) - degree, t)
-            - std::begin(knots) - 1;
-        return span;
+            std::end(knots) - degree, t) - std::begin(knots) - 1;
+        return static_cast<int>(span);
     }
 };
